hj: do not install rvm if exists

+++ test.sh
@@ -1,9 +1,15 @@
 #!/bin/bash
+
 echo '----prepare----'
-sudo yum update
-sudo yum install curl
-curl -L get.rvm.io | bash -s stable
-source ~/.rvm/rvm.sh
+if [[ -s "$HOME/.rvm/bin/rvm" ]] ; then
+  source "$HOME/.rvm/bin/rvm"
+else
+  sudo yum update
+  sudo yum install curl
+  curl -L get.rvm.io | bash -s stable
+  source ~/.rvm/bin/rvm
+fi
+
 
 # rvm --default use 1.9.3
 # bundle install
